// src/organizations/dto/organizations.output.dto.ts
export class OrganizationOutputDto {
	id: number;
	name: string;
	category: string;
	description: string;
	country: string;
	state: string;
	city: string;
	website: string;
	phone: string;
	adress: string;
	zip: string;
	logo: string;
	backgroundImage: string;
	terms: boolean;
	// specialUsers: [];
}